correcciones

+++ backend/index.js
@@ -2,6 +2,7 @@ const express = require('express');
 const cors = require('cors');
 const app = express();
 const { Pool } = require('pg');
+const { verify, Verify } = require('crypto');
 const router = express.Router();
 require('dotenv').config();
 const { key, PORT } = process.env;
@@ -45,7 +46,9 @@ router.post('/posts', async (req, res) => {
       const { titulo, url, descripcion } = req.body;
 
       if (!titulo || !url || !descripcion) {
-        return res.status(400).json({ error: 'Faltan uno o más campos requeridos' });
+        return (res.status(400).json({ error: 'Faltan uno o más campos requeridos' }),
+                console.log("error, Faltan uno o más campos requeridos")
+      );
       }
 
       const queryText = `
@@ -64,17 +67,39 @@ router.post('/posts', async (req, res) => {
 
   //de la segunda evaluacion:
 
+
+  // correccion de lo mencionado en la evaluacion, la siguiente funcion se encarga de validar el id ingresado:
+    const verifyId = async (idForVerify) => {
+    const getResponse = await pool.query(`SELECT * FROM posts; `);
+    const all = getResponse.rows;
+    const verificated = all.find(all => all.id == idForVerify);
+    console.log(verificated);
+    if(verificated){
+      return (verificated.id);
+    } else {
+      console.log("error, id no encontrado");
+      return
+    }
+  };
+
 router.put('/posts/like/:id', async (req, res) => {
   try {
     const idModificar = req.params.id;
-    console.log('el id a modificar es', idModificar )
+    const idVerificated = await verifyId(idModificar);
+    console.log("mi id verificated es", idVerificated);
 
-    const queryText = `
+    if(idModificar == idVerificated){
+      const queryText = `
       UPDATE posts SET likes = likes + 1
       WHERE ID = $1;
     `;
-    const getResponse = await pool.query(queryText, [idModificar]);
-    res.status(201).json(getResponse);
+    console.log('el id a modificar sigue siendo', idModificar );
+    await pool.query(queryText, [idModificar]);
+    const getResponse = await pool.query(`SELECT * FROM posts WHERE ID = $1; `, [idModificar] );
+    res.send(`like agregado en item: ${JSON.stringify(getResponse.rows)}`);
+    } else {
+      res.send(`el id ${idModificar} no es valido`);
+    }
   } catch (error) {
     console.error('Error al modificar datos:', error);
     res.status(500).json({ error: 'Error en el servidor' });
@@ -85,21 +110,24 @@ router.put('/posts/like/:id', async (req, res) => {
 router.delete('/posts/:id', async (req, res) => {
   try {
     const idBorrar = req.params.id;
-    console.log('el id a borrar es', idBorrar )
+    const idVerificated = await verifyId(idBorrar);
+    console.log("mi id verificated es", idVerificated);
 
+  if(idBorrar == idVerificated){
     const queryText = `
-      DELETE FROM posts
-      WHERE ID = $1;
-    `;
-    const getResponse = await pool.query(queryText, [idBorrar]);
-    res.status(201).json(getResponse);
+    DELETE FROM posts
+    WHERE ID = $1;
+  `;
+  await pool.query(queryText, [idBorrar]);
+  res.send(`item con id ${idBorrar} eliminado con exito`);
+  } else {
+    res.send(`el id ${idBorrar} no es valido`);
+  }
   } catch (error) {
     console.error('Error al borrar datos:', error);
     res.status(500).json({ error: 'Error en el servidor' });
   }
 });
-
-
 
 app.listen(PORT, async () => {
   console.log(`Server is running on port ${PORT}`);
